# server/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "生产级 DDD 脚手架项目的 API 文档，包含用户认证、事件驱动等核心功能。",
        "title": "Go DDD Scaffold API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.github.com/shenfay/kiqi",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "管理员查询所有消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "消息分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息列表",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "获取当前用户消息列表",
                "parameters": [
                    {
                        "enum": [
                            "system",
                            "companion"
                        ],
                        "type": "string",
                        "description": "消息类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "消息分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "已读状态",
                        "name": "is_read",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息列表",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/read-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "标记全部已读",
                "parameters": [
                    {
                        "description": "可选类型筛选",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_app_notification.ReadAllCmd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "标记成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "获取未读消息计数",
                "responses": {
                    "200": {
                        "description": "未读计数",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/{id}/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "标记消息已读",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "标记成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "消息不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_shenfay_kiqi_internal_app_notification.ReadAllCmd": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "响应时间（RFC3339 格式）",
                    "type": "string"
                },
                "trace_id": {
                    "description": "链路追踪 ID（用于日志关联和分布式追踪）",
                    "type": "string"
                }
            }
        },
        "github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "响应时间（RFC3339 格式）",
                    "type": "string"
                },
                "trace_id": {
                    "description": "链路追踪 ID（用于日志关联和分布式追踪）",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 JWT Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}